CArtifactHolder divided into files
